// powerapps-project-planner-20260612/src/generated/models/Vibe_projectsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Vibe_projectsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Vibe_projectsstatecode = keyof typeof Vibe_projectsstatecode;
export const Vibe_projectsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Vibe_projectsstatuscode = keyof typeof Vibe_projectsstatuscode;

export interface Vibe_projectsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Vibe_projectsstatecode;
  statuscode?: Vibe_projectsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  vibe_budgetremaining?: number;
  vibe_budgetspent?: number;
  "vibe_clientid@odata.bind"?: string;
  vibe_description?: string;
  vibe_enddate?: string;
  vibe_name?: string;
  "vibe_ownerid@odata.bind"?: string;
  vibe_projectid: string;
  vibe_sharepointfolderurl?: string;
  vibe_startdate?: string;
  "vibe_templatesourceid@odata.bind"?: string;
  vibe_totalbudget?: number;
}

export interface Vibe_projects extends Vibe_projectsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  vibe_clientidname?: string;
  vibe_clientidyominame?: string;
  vibe_owneridname?: string;
  vibe_owneridyominame?: string;
  vibe_templatesourceidname?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  vibe_clientid?: object;
  _vibe_clientid_value?: string;
  vibe_ownerid?: object;
  _vibe_ownerid_value?: string;
  vibe_templatesourceid?: object;
  _vibe_templatesourceid_value?: string;
}