// powerapps-project-planner-20260612/src/generated/models/Vibe_projectrolesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Vibe_projectrolesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Vibe_projectrolesstatecode = keyof typeof Vibe_projectrolesstatecode;
export const Vibe_projectrolesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Vibe_projectrolesstatuscode = keyof typeof Vibe_projectrolesstatuscode;

export interface Vibe_projectrolesBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Vibe_projectrolesstatecode;
  statuscode?: Vibe_projectrolesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  vibe_defaulthourlyrate?: number;
  vibe_description?: string;
  vibe_isstakeholderrole?: boolean;
  vibe_name?: string;
  vibe_projectroleid: string;
}

export interface Vibe_projectroles extends Vibe_projectrolesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  vibe_isstakeholderrolename?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}